requirements.txt
flask==3.0.2
flask-cors==4.0.0
psycopg2-binary==2.9.9
redis==5.0.1
boto3==1.34.69
python-dotenv==1.0.1
gunicorn==21.2.0
sqlalchemy==2.0.28
